add DataFlowOverlay, fix stylings, refactor
