+++ web/src/components/ProcessTimeline2/TimelinePerUnit.scss
@@ -1,0 +1,120 @@
+.timeline-per-unit {
+  width: 100%;
+  height: 100%;
+  background-color: white;
+  border: 1px solid #dee2e6;
+  border-radius: 6px;
+  position: relative;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+  font-family: monospace;
+  overflow-x: auto;
+  overflow-y: hidden;
+
+  .units-header {
+    padding: 1rem;
+    font-weight: bold;
+    border-bottom: 1px solid #dee2e6;
+    background-color: #f8f9fa;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: sticky;
+    top: 0;
+    z-index: 10;
+  }
+
+  .units-container {
+    position: relative;
+    padding: 0 1rem;
+    display: flex;
+    flex-direction: row;
+    gap: 1rem;
+    min-height: 100%;
+  }
+
+  .unit-row {
+    flex: 1;
+    min-width: 80px;
+    display: flex;
+    flex-direction: column;
+    //border-right: 1px dashed #dee2e6;
+
+    &:last-child {
+      border-right: none;
+    }
+
+    .unit-name {
+      padding: 0.5rem;
+      font-size: 0.8rem;
+      font-weight: bold;
+      text-align: center;
+      word-break: break-all;
+      background-color: #f8f9fa;
+      border-bottom: 1px solid #dee2e6;
+    }
+
+    .unit-timeline-track {
+      position: relative;
+      flex: 1;
+      min-height: 1000px;
+
+      .function-rectangle {
+        position: absolute;
+        z-index: 1;
+        border-radius: 6px;
+        box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+        padding: 0;
+
+        .function-header {
+          display: none;
+        }
+
+        .instructions-container {
+          position: relative;
+
+          .instruction-rectangle {
+            z-index: 3;
+            position: absolute;
+            left: 0.5rem;
+            right: 0.5rem;
+            min-height: 50px;
+            box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+            cursor: pointer;
+            display: flex;
+            align-items: center;
+            text-align: center;
+            background: white;
+            border-radius: 6px;
+
+            &:hover {
+              box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
+            }
+
+            .instruction-content {
+              width: 100%;
+              display: flex;
+              flex-direction: column;
+              line-height: 1.2;
+
+              .instruction-label {
+                font-size: 0.8rem;
+                color: #2d3748;
+                font-family: Calibri;
+              }
+
+              .instruction-time {
+                font-size: 0.7rem;
+                color: #6c757d;
+              }
+
+              .instruction-io-info {
+                font-size: 0.65rem;
+                color: #6c757d;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
