--- web/src/components/TimelinePerUnit.scss
@@ -1,84 +1,0 @@
-.timeline-per-unit {
-  width: 100%;
-  height: 100%;
-  background-color: white;
-  border: 1px solid #dee2e6;
-  border-radius: 6px;
-  position: relative;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-  font-family: monospace;
-  overflow-x: auto;
-  overflow-y: hidden;
-
-  .units-header {
-    padding: 1rem;
-    font-weight: bold;
-    border-bottom: 1px solid #dee2e6;
-    background-color: #f8f9fa;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: sticky;
-    top: 0;
-    z-index: 10;
-  }
-
-  .units-container {
-    position: relative;
-    padding: 0 1rem;
-    display: flex;
-    flex-direction: row;
-    gap: 1rem;
-    min-height: 100%;
-  }
-
-  .unit-row {
-    flex: 1;
-    min-width: 80px;
-    display: flex;
-    flex-direction: column;
-    border-right: 1px dashed #dee2e6;
-
-    &:last-child {
-      border-right: none;
-    }
-
-    .unit-name {
-      padding: 0.5rem;
-      font-size: 0.8rem;
-      font-weight: bold;
-      text-align: center;
-      word-break: break-all;
-      background-color: #f8f9fa;
-      border-bottom: 1px solid #dee2e6;
-    }
-
-    .unit-timeline-track {
-      position: relative;
-      flex: 1;
-      min-height: 1000px; // Should probably match containerHeight or use absolute positioning based on time
-    }
-
-    .unit-function-block {
-      position: absolute;
-      left: 10%;
-      width: 80%;
-      background-color: rgba(0, 123, 255, 0.2);
-      border: 1px solid rgba(0, 123, 255, 0.5);
-      border-radius: 4px;
-      transition: background-color 0.2s;
-
-      &:hover {
-        background-color: rgba(0, 123, 255, 0.4);
-      }
-
-      .unit-block-label {
-        font-size: 0.65rem;
-        padding: 2px;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-      }
-    }
-  }
-}
